extraction/google-drive/requirements.txt
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1